% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_tracker.R
\name{run_tracker}
\alias{run_tracker}
\title{Eu covid19 tracker}
\usage{
run_tracker()
}
\value{
a shiny app for tracking covid19 in European countries
}
\description{
Function launching the application
}
\author{
Aëllya & Laurène
}
